requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
Django==3.0.8
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-filters==0.11.1
filelock==3.0.12
flake8==3.8.3
httpie==2.2.0
idna==2.10
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
jupyter-client==6.1.5
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.2
mccabe==0.6.1
numpy==1.19.0
parso==0.7.0
pbr==5.4.5
pickleshare==0.7.5
pipenv==2020.6.2
prompt-toolkit==3.0.5
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pyzmq==19.0.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stevedore==3.2.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
virtualenv==20.0.28
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
wrapt==1.12.1
